segment schwierigkeitsgrad

+++ FitWithJumpAndRun/GameLogic/LevelSegment.cs
@@ -22,7 +22,7 @@ namespace JumpAndRun.GameLogic
         /// <summary>Länge des Levelsegmentes</summary>
         public float Length { get; set; }
         /// <summary>Schwierigkeitsgrad des Levelsegmentes</summary>
-        public float Severity { get; set; }
+        public int Severity { get; set; }
 
         /// <summary>
         /// Initialisierung des Levelsegments
